Refactor quality monitoring modules
